pyproject.toml
[tool.poetry]
name = "nonebot-plugin-disconnect-notice"
version = "0.4.0"
description = "bot断连时的通知插件"
authors = ["cypas <ayano05@outlook.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_disconnect_notice"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2="^2.2.0"
nonebot-adapter-onebot = "^2.4.1"
aiosmtplib = "^2.0.2"
httpx="^0.24.1"
nonebot-plugin-apscheduler = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
